# docs/swagger.yaml
basePath: /
definitions:
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  domain.MarketingService:
    properties:
      benefits:
        items:
          type: string
        type: array
      description:
        type: string
      features:
        items:
          type: string
        type: array
      icon_url:
        type: string
      id:
        type: integer
      marketing_name:
        type: string
      screenshot_url:
        type: string
      tag_line:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  domain.RefreshTokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  domain.SuccessResponse:
    properties:
      data:
        description: Omits Data field if nil
      message:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: suporte@solude.tech
    name: Eng. Gabriel Coelho | InovaData development team
    url: https://solude.tech
  description: Platform API is a RESTful API for managing ...
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Platform API
  version: 0.1.1
paths:
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Sends an email to the user with a link to reset their password
      operationId: forgotPassword
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email sent successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request - Invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Forgot Password
      tags:
      - Auth User
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using their email and password, then returns
        access and refresh tokens for session management.
      operationId: login
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login, returns access and refresh tokens
          schema:
            $ref: '#/definitions/domain.LoginResponse'
        "400":
          description: Bad Request - Invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized - Incorrect email or password
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Login user
      tags:
      - Auth User
  /login-guest:
    post:
      consumes:
      - application/json
      description: Authenticates a guest user using their IP address, then returns
        access and refresh tokens for session management
      operationId: loginGuest
      produces:
      - application/json
      responses:
        "200":
          description: Successful login, returns access and refresh tokens
          schema:
            $ref: '#/definitions/domain.LoginResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Login Guest
      tags:
      - Auth User
  /refresh-token:
    post:
      consumes:
      - application/json
      description: Refreshes the user's access token
      operationId: refreshToken
      parameters:
      - description: Refresh token
        in: query
        name: refreshToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Access token refreshed successfully
          schema:
            $ref: '#/definitions/domain.RefreshTokenResponse'
        "400":
          description: Bad Request - Invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized - Invalid refresh token
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Refresh Token
      tags:
      - Auth User
  /reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password
      operationId: resetPassword
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      - description: New password
        in: query
        name: newPassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request - Invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Reset Password
      tags:
      - Auth User
...
  /services/marketing:
    get:
      description: Gets all marketing services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.MarketingService'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get Marketing Services
      tags:
      - Website Service
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"